requirements.txt
webrtcvad
pyAudioAnalysis
openai-whisper
pyaudio 
